third_party/WebKit/LayoutTests/fast/canvas-api/canvas-overloads-drawImage.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<script>
test(function(t) {

    var ctx = document.createElement('canvas').getContext('2d');
    
    var imageElement = document.createElement("img");
    try {
      ctx.drawImage();
    } catch (ex) {
      assert_equals(ex.name, "TypeError");
    }
    try {
      ctx.drawImage(imageElement);
    } catch (ex) {
      assert_equals(ex.name, "TypeError");
    }
    try {
      ctx.drawImage(imageElement, 0);
    } catch (ex) {
      assert_equals(ex.name, "TypeError");
    }
    assert_equals(ctx.drawImage(imageElement, 0, 0), undefined);
    try {
      ctx.drawImage(imageElement, 0, 0, 0);
    } catch (ex) {
      assert_equals(ex.name, "TypeError");
    }
    assert_equals(ctx.drawImage(imageElement, 0, 0, 0, 0), undefined);
    try {
      ctx.drawImage(imageElement, 0, 0, 0, 0, 0);
    } catch (ex) {
      assert_equals(ex.name, "TypeError");
    }
    try {
      ctx.drawImage(imageElement, 0, 0, 0, 0, 0, 0);
    } catch (ex) {
      assert_equals(ex.name, "TypeError");
    }
    try {
      ctx.drawImage(imageElement, 0, 0, 0, 0, 0, 0, 0);
    } catch (ex) {
      assert_equals(ex.name, "TypeError");
    }
    assert_equals(ctx.drawImage(imageElement, 0, 0, 0, 0, 0, 0, 0, 0), undefined);
    assert_equals(ctx.drawImage(imageElement, 0, 0, 0, 0, 0, 0, 0, 0, 0), undefined);
    assert_equals(ctx.drawImage(imageElement, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0), undefined);
    
    var canvasElement = document.createElement("canvas");
    try {
      ctx.drawImage(canvasElement);
    } catch (ex) {
      assert_equals(ex.name, "TypeError");
    }
    try {
      ctx.drawImage(canvasElement, 0);
    } catch (ex) {
      assert_equals(ex.name, "TypeError");
    }
    assert_equals(ctx.drawImage(canvasElement, 0, 0), undefined);
    try {
      ctx.drawImage(canvasElement, 0, 0, 0);
    } catch (ex) {
      assert_equals(ex.name, "TypeError");
    }
    assert_equals(ctx.drawImage(canvasElement, 0, 0, 0, 0), undefined);
    try {
      ctx.drawImage(canvasElement, 0, 0, 0, 0, 0);
    } catch (ex) {
      assert_equals(ex.name, "TypeError");
    }
    try {
      ctx.drawImage(canvasElement, 0, 0, 0, 0, 0, 0);
    } catch (ex) {
      assert_equals(ex.name, "TypeError");
    }
    try {
      ctx.drawImage(canvasElement, 0, 0, 0, 0, 0, 0, 0);
    } catch (ex) {
      assert_equals(ex.name, "TypeError");
    }
    assert_equals(ctx.drawImage(canvasElement, 0, 0, 0, 0, 0, 0, 0, 0), undefined);
    assert_equals(ctx.drawImage(canvasElement, 0, 0, 0, 0, 0, 0, 0, 0, 0), undefined);
    assert_equals(ctx.drawImage(canvasElement, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0), undefined);

}, "Test the behavior of CanvasRenderingContext2D.drawImage() when called with different numbers of arguments.");
</script>
</body>
